Extract out message handlers to a partial class
