+++ Libplanet/Net/Swarm.cs
@@ -21,7 +21,6 @@ using Libplanet.Store;
 using Libplanet.Tx;
 using Nito.AsyncEx;
 using Serilog;
-using Serilog.Events;
 
 namespace Libplanet.Net
 {
@@ -1595,171 +1594,11 @@ namespace Libplanet.Net
             BroadcastMessage(except, message);
         }
 
-        private void ProcessMessageHandler(object target, Message message)
-        {
-            switch (message)
-            {
-                case Ping ping:
-                    {
-                        _logger.Debug($"Ping received.");
-
-                        // This case can be dealt in Transport.
-                        var pong = new Pong()
-                        {
-                            Identity = ping.Identity,
-                        };
-
-                        Transport.ReplyMessage(pong);
-                        break;
-                    }
-
-                case GetChainStatus getChainStatus:
-                    {
-                        _logger.Debug($"GetChainStatus received.");
-
-                        // This is based on the assumption that genesis block always exists.
-                        var chainStatus = new ChainStatus(
-                            BlockChain.Tip.Index,
-                            BlockChain.Tip.TotalDifficulty)
-                        {
-                            Identity = getChainStatus.Identity,
-                        };
-
-                        Transport.ReplyMessage(chainStatus);
-                        break;
-                    }
-
-                case FindNeighbors findPeer:
-                    {
-                        _logger.Debug($"FindNeighbors received.");
-                        break;
-                    }
-
-                case GetBlockHashes getBlockHashes:
-                    {
-                        BlockChain.FindNextHashes(
-                            getBlockHashes.Locator,
-                            getBlockHashes.Stop,
-                            FindNextHashesChunkSize
-                        ).Deconstruct(
-                            out long? offset,
-                            out IReadOnlyList<HashDigest<SHA256>> hashes
-                        );
-                        var reply = new BlockHashes(offset, hashes)
-                        {
-                            Identity = getBlockHashes.Identity,
-                        };
-                        Transport.ReplyMessage(reply);
-                        break;
-                    }
-
-                case GetRecentStates getRecentStates:
-                    {
-                        TransferRecentStates(getRecentStates);
-                        break;
-                    }
-
-                case GetBlocks getBlocks:
-                    {
-                        TransferBlocks(getBlocks);
-                        break;
-                    }
-
-                case GetTxs getTxs:
-                    {
-                        TransferTxs(getTxs);
-                        break;
-                    }
-
-                case TxIds txIds:
-                    {
-                        ProcessTxIds(txIds);
-                        break;
-                    }
-
-                case BlockHashes blockHashes:
-                    {
-                        _logger.Error("BlockHashes message is only for IBD.");
-                        break;
-                    }
-
-                case BlockHeaderMessage blockHeader:
-                    {
-                        Task.Run(
-                            async () => await ProcessBlockHeader(blockHeader, _cancellationToken),
-                            _cancellationToken);
-                        break;
-                    }
-
-                case GetBlockStates getBlockStates:
-                    TransferBlockStates(getBlockStates);
-                    break;
-
-                default:
-                    throw new InvalidMessageException($"Can't handle message: {message}", message);
-            }
-        }
-
         private bool IsDemandNeeded(BlockHeader target)
         {
             return target.TotalDifficulty > BlockChain.Tip.TotalDifficulty &&
                    (_demandBlockHash is null ||
                     _demandBlockHash.Value.Header.TotalDifficulty < target.TotalDifficulty);
-        }
-
-        private async Task ProcessBlockHeader(
-            BlockHeaderMessage message,
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            if (!(message.Remote is BoundPeer peer))
-            {
-                _logger.Information(
-                    "BlockHeaderMessage was sent from invalid peer " +
-                    "{PeerAddress}; ignored.",
-                    message.Remote.Address);
-                return;
-            }
-
-            BlockHeaderReceived.Set();
-            BlockHeader header = message.Header;
-
-            _logger.Debug(
-                "Block header of hash [{Hash}] (index: {Index}) received.",
-                ByteUtil.Hex(header.Hash),
-                header.Index,
-                Address.ToHex());
-
-            try
-            {
-                header.Validate(DateTimeOffset.UtcNow);
-            }
-            catch (InvalidBlockException ibe)
-            {
-                _logger.Information(
-                    ibe,
-                    "Received header[{Hash}] seems invalid; ignore.",
-                    ByteUtil.Hex(header.Hash)
-                );
-                return;
-            }
-
-            using (await _blockSyncMutex.LockAsync(cancellationToken))
-            {
-                if (IsDemandNeeded(header))
-                {
-                    _demandBlockHash = new BlockHashDemand(header, peer);
-                }
-                else
-                {
-                    _logger.Debug(
-                        "No blocks are required " +
-                        "(current: {Current}, demand: {Demand}, received: {Received});" +
-                        $" {nameof(BlockHeaderMessage)} is ignored.",
-                        BlockChain.Tip.Index,
-                        _demandBlockHash?.Header.Index,
-                        header.Index);
-                }
-            }
         }
 
         private async Task SyncPreviousBlocksAsync(
@@ -1995,243 +1834,6 @@ namespace Libplanet.Net
             }
 
             return workspace;
-        }
-
-        private void TransferTxs(GetTxs getTxs)
-        {
-            IEnumerable<Transaction<T>> txs = getTxs.TxIds
-                .Where(txId => _store.ContainsTransaction(txId))
-                .Select(BlockChain.GetTransaction);
-
-            foreach (Transaction<T> tx in txs)
-            {
-                Message response = new Messages.Tx(tx.Serialize(true))
-                {
-                    Identity = getTxs.Identity,
-                };
-                Transport.ReplyMessage(response);
-            }
-        }
-
-        private void ProcessTxIds(TxIds message)
-        {
-            if (!(message.Remote is BoundPeer peer))
-            {
-                _logger.Information(
-                    "TxIds was sent from invalid peer " +
-                    $"[{message.Remote?.Address.ToHex()}]. ignored.");
-                return;
-            }
-
-            _logger.Debug(
-                "Received TxIds: [{txIds}]",
-                string.Join(", ", message.Ids));
-
-            ImmutableHashSet<TxId> newTxIds = message.Ids
-                .Where(id => !_demandTxIds.ContainsKey(id))
-                .Where(id => !_store.ContainsTransaction(id))
-                .ToImmutableHashSet();
-
-            if (!newTxIds.Any())
-            {
-                _logger.Debug("No txs to require.");
-                return;
-            }
-
-            _logger.Debug("Txs to require: {@txIds}", newTxIds.Select(txid => txid.ToString()));
-            foreach (var txid in newTxIds)
-            {
-                _demandTxIds.TryAdd(txid, peer);
-            }
-        }
-
-        private void TransferBlocks(GetBlocks getData)
-        {
-            string identityHex = ByteUtil.Hex(getData.Identity);
-            _logger.Verbose("Preparing a blocks reply to request {Identity}...", identityHex);
-
-            var blocks = new List<byte[]>();
-
-            List<HashDigest<SHA256>> hashes = getData.BlockHashes.ToList();
-            int i = 1;
-            int total = hashes.Count;
-            const string logMsg =
-                "Fetching a block #{Index}/{Total} ({Hash}) to include to " +
-                "a reply to {Identity}...";
-            foreach (HashDigest<SHA256> hash in hashes)
-            {
-                _logger.Verbose(logMsg, i, total, hash, identityHex);
-                if (_store.ContainsBlock(hash))
-                {
-                    Block<T> block = _store.GetBlock<T>(hash);
-                    byte[] payload = block.Serialize();
-                    blocks.Add(payload);
-                }
-
-                if (blocks.Count == getData.ChunkSize)
-                {
-                    var response = new Messages.Blocks(blocks)
-                    {
-                        Identity = getData.Identity,
-                    };
-                    _logger.Verbose(
-                        "Enqueuing a blocks reply (...{Index}/{Total})...",
-                        i,
-                        total
-                    );
-                    Transport.ReplyMessage(response);
-                    blocks.Clear();
-                }
-
-                i++;
-            }
-
-            if (blocks.Any())
-            {
-                var response = new Messages.Blocks(blocks)
-                {
-                    Identity = getData.Identity,
-                };
-                _logger.Verbose(
-                    "Enqueuing a blocks reply (...{Index}/{Total}) to {Identity}...",
-                    total,
-                    total,
-                    identityHex
-                );
-                Transport.ReplyMessage(response);
-            }
-
-            _logger.Debug("Blocks were transferred to {Identity}.", identityHex);
-        }
-
-        private void TransferRecentStates(GetRecentStates getRecentStates)
-        {
-            BlockLocator baseLocator = getRecentStates.BaseLocator;
-            HashDigest<SHA256>? @base = BlockChain.FindBranchPoint(baseLocator);
-            HashDigest<SHA256> target = getRecentStates.TargetBlockHash;
-            IImmutableDictionary<HashDigest<SHA256>,
-                IImmutableDictionary<string, IValue>
-            > blockStates = null;
-            IImmutableDictionary<string, IImmutableList<HashDigest<SHA256>>> stateRefs = null;
-            long nextOffset = -1;
-            int iteration = 0;
-
-            if (BlockChain.ContainsBlock(target))
-            {
-                ReaderWriterLockSlim rwlock = BlockChain._rwlock;
-                rwlock.EnterReadLock();
-                try
-                {
-                    Guid chainId = BlockChain.Id;
-
-                    _logger.Debug(
-                        "Getting state references from {Offset}",
-                        getRecentStates.Offset);
-
-                    long baseIndex =
-                        (@base is HashDigest<SHA256> bbh &&
-                         _store.GetBlockIndex(bbh) is long bbIdx)
-                            ? bbIdx
-                            : 0;
-                    long lowestIndex = baseIndex + getRecentStates.Offset;
-                    long targetIndex =
-                        (target is HashDigest<SHA256> tgt &&
-                         _store.GetBlockIndex(tgt) is long tgtIdx)
-                            ? tgtIdx
-                            : long.MaxValue;
-
-                    iteration =
-                        (int)Math.Ceiling(
-                            (double)(targetIndex - baseIndex + 1) / FindNextStatesChunkSize);
-                    _logger.Verbose("Iteration is : {Iteration}", iteration);
-
-                    long highestIndex = lowestIndex + FindNextStatesChunkSize - 1 > targetIndex
-                        ? targetIndex
-                        : lowestIndex + FindNextStatesChunkSize - 1;
-
-                    nextOffset = highestIndex == targetIndex
-                        ? -1
-                        : getRecentStates.Offset + FindNextStatesChunkSize;
-
-                    stateRefs = _store.ListAllStateReferences(
-                        chainId,
-                        lowestIndex: lowestIndex,
-                        highestIndex: highestIndex
-                    );
-                    if (_logger.IsEnabled(LogEventLevel.Verbose))
-                    {
-                        _logger.Verbose(
-                            "List state references from {From} to {To}:\n{StateReferences}",
-                            lowestIndex,
-                            highestIndex,
-                            string.Join(
-                                "\n",
-                                stateRefs.Select(kv => $"{kv.Key}: {string.Join(", ", kv.Value)}")
-                            )
-                        );
-                    }
-
-                    // GetBlockStates may return null since swarm may not have deep states.
-                    blockStates = stateRefs.Values
-                        .Select(refs => refs.Last())
-                        .ToImmutableHashSet()
-                        .Select(bh => (bh, _store.GetBlockStates(bh)))
-                        .Where(pair => !(pair.Item2 is null))
-                        .ToImmutableDictionary(
-                            pair => pair.Item1,
-                            pair => (IImmutableDictionary<string, IValue>)pair.Item2
-                                .ToImmutableDictionary(kv => kv.Key, kv => kv.Value)
-                        );
-                }
-                finally
-                {
-                    rwlock.ExitReadLock();
-                }
-            }
-
-            if (_logger.IsEnabled(LogEventLevel.Verbose))
-            {
-                if (BlockChain.ContainsBlock(target))
-                {
-                    var baseString = @base is HashDigest<SHA256> h
-                        ? $"{BlockChain[h].Index}:{h}"
-                        : null;
-                    var targetString = $"{BlockChain[target].Index}:{target}";
-                    _logger.Verbose(
-                        "State references to send (preload): {StateReferences} ({Base}-{Target})",
-                        stateRefs.Select(kv =>
-                            (
-                                kv.Key,
-                                string.Join(", ", kv.Value.Select(v => v.ToString()))
-                            )
-                        ).ToArray(),
-                        baseString,
-                        targetString
-                    );
-                    _logger.Verbose(
-                        "Block states to send (preload): {BlockStates} ({Base}-{Target})",
-                        blockStates.Select(kv => (kv.Key, kv.Value)).ToArray(),
-                        baseString,
-                        targetString
-                    );
-                }
-                else
-                {
-                    _logger.Verbose("Nothing to reply because {TargetHash} doesn't exist.", target);
-                }
-            }
-
-            var reply = new RecentStates(
-                target,
-                nextOffset,
-                iteration,
-                blockStates,
-                stateRefs?.ToImmutableDictionary())
-            {
-                Identity = getRecentStates.Identity,
-            };
-
-            Transport.ReplyMessage(reply);
         }
 
         private async Task ProcessFillblock(CancellationToken cancellationToken)
